requirements.txt
anyio==3.6.2
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
black==23.1.0
click==8.1.3
decorator==5.1.1
executing==1.2.0
fastapi==0.92.0
h11==0.14.0
idna==3.4
ipython==8.10.0
jedi==0.18.2
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==23.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.0.0
prompt-toolkit==3.0.37
psycopg==3.1.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.5
Pygments==2.14.0
pyhumps==3.8.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
stack-data==0.6.2
starlette==0.25.0
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.5.0
uvicorn==0.20.0
wcwidth==0.2.6
anyio==3.6.2
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
black==23.1.0
click==8.1.3
decorator==5.1.1
executing==1.2.0
fastapi==0.92.0
h11==0.14.0
idna==3.4
ipython==8.10.0
jedi==0.18.2
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
packaging==23.0
parso==0.8.3
pathspec==0.11.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.0.0
prompt-toolkit==3.0.37
psycopg==3.1.8
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.5
Pygments==2.14.0
pyhumps==3.8.0
python-dotenv==1.0.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
stack-data==0.6.2
starlette==0.25.0
tomli==2.0.1
traitlets==5.9.0
typing_extensions==4.5.0
uvicorn==0.20.0
wcwidth==0.2.6
